+++ tests/test_cli.py
@@ -5,6 +5,7 @@ from click.testing import CliRunner
 import pytest
 
 from pkgmt import cli
+from invoke import Context
 
 
 @pytest.mark.parametrize(
@@ -262,3 +263,26 @@ def test_lint_black_pyproj(pyproject, command, output, tmp_empty):
     result = runner.invoke(cli.cli, command)
 
     assert output in result.output
+
+
+@pytest.mark.parametrize(
+    "command, version, doc",
+    [
+        [["setup"], None, False],
+        [["setup", "--version", "2.7"], "2.7", False],
+        [["setup", "--version", "3.7"], "3.7", False],
+        [["setup", "--version", "3.8"], "3.8", False],
+        [["setup", "--version", "3.9"], "3.9", False],
+        [["setup", "--version", "3.10"], "3.10", False],
+        [["setup", "--version", "3.11"], "3.11", False],
+        [["setup", "--version", "3.9", "--doc"], "3.9", True],
+    ],
+)
+def test_setup_with_python_version(monkeypatch, command, version, doc):
+    mock = Mock()
+    monkeypatch.setattr(cli.dev, "setup", mock)
+
+    runner = CliRunner()
+    runner.invoke(cli.cli, command)
+
+    mock.assert_called_once_with(Context(), version=version, doc=doc)